+++ Source/Algorithm/Twofish/Twofish.cpp
@@ -43,7 +43,7 @@ String Twofish::Encrypt(const String& password, const String& buffer)
 	const String iv = Algorithm::GenerateRandomIV(CryptoPP::Twofish::BLOCKSIZE); 
 	
 	// Prepare key
-	CryptoPP::SecByteBlock key(CryptoPP::Twofish::MAX_KEYLENGTH + CryptoPP::Twofish::BLOCKSIZE); // Encryption key to be generated from user password + IV
+	CryptoPP::SecByteBlock key(static_cast<std::size_t>(CryptoPP::Twofish::MAX_KEYLENGTH) + static_cast<std::size_t>(CryptoPP::Twofish::BLOCKSIZE)); // Encryption key to be generated from user password + IV
 
 	// Convert key to KDF SHA-256, which allows you to use a password smaller or larger than the cipher's key size
 	CryptoPP::HKDF<CryptoPP::SHA256> hkdf{};
